+++ api/apiDefinitions/verify.js
@@ -10,16 +10,17 @@ verify.use(bodyParser.json());
 //APIs
 
 verify.get("/", async (req, res) => {
+  console.log("req.headers", req.headers);
   const authHeader = req.headers.authorization || req.headers.Authorization;
   if (!authHeader?.startsWith("Bearer ")) return res.sendStatus(401);
   const token = authHeader.split(" ")[1];
 
   jwt.verify(token, process.env.ACCESS_TOKEN_SECRET, (err, decoded) => {
     console.log("verify.js", "err", err, "decoded", decoded);
-    if (err) return res.status(403).json(false); //invalid token
+    if (err) return res.sendStatus(403); //invalid token
     res.user = decoded.email;
 
-    return res.status(200).json(true);
+    return res.sendStatus(200);
   });
 });
