+++ db/users/u_d814afd9-2afd-47a0-a91a-d051a928bf3d.json
@@ -1,8 +1,8 @@
 {
   "created": "Tue Jan 25 2022 21:05:55 GMT+0100 (centraleuropeisk normaltid)",
-  "updated": "Fri Jan 28 2022 01:41:28 GMT+0100 (centraleuropeisk normaltid)",
+  "updated": "Fri Jan 28 2022 21:51:37 GMT+0100 (centraleuropeisk normaltid)",
   "email": "nicholas@nikdev.io",
   "hashedPwd": "$2b$10$hFoAYFN0xhcXXqL27mbfeuETESQ9wqeajcQscX0GAdSOH51fzWEJG",
   "id": "u_d814afd9-2afd-47a0-a91a-d051a928bf3d",
-  "refreshToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6Im5pY2hvbGFzQG5pa2Rldi5pbyIsImlhdCI6MTY0MzMzMDQ4OCwiZXhwIjoxNjQzNDE2ODg4fQ.EqwbWKWOoE8Jsn3MR65zGQHgaO9X6ddAjS5W_hKi6Is"
+  "refreshToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6Im5pY2hvbGFzQG5pa2Rldi5pbyIsImlhdCI6MTY0MzQwMzA5NywiZXhwIjoxNjQzNDg5NDk3fQ.OoP42v2NKEXrFP1pNyvkQuZOpNhkTBfMJ9cX-fE4n1U"
 }
